Joe's code
This upload includes files needed for the diet composition analysis in relation to morphology and sites, using a GLMM. It also includes code for producing the map. 

The file "dissection_data.csv" is from Ragna, and appears to be identical to "Final.csv" (I simply renamed it). So, it might make sense to include just one version of this file, modifying code accordingly. This file, along with "pca_scores.csv" (which was previously uploaded and I assume is the same one I used) and "stations.csv" are the only ones needed to run my code. 

My code produces two csv files: "diet_prevalence.csv" and "diet_long_analysis.csv", which I have uploaded. The latter is needed to run the R script "Fit Taxon-Specific Models.R"